refactor(db): consolidate packages into single db package
the db definitions/schemas/queries for both the api and consumer project
has been consolidated into a single db package that serves both projects
(i.e. both the consumer/subscriber and web api).

this was done to reduce duplicated code and to hopefully just make
managing the project as a whole better.